+++ releases.md
@@ -4,6 +4,7 @@
 
   - Start health checker earlier in the process. Use `#healthy!` message instead of `#ready!`.
   - Emit prepared and running log messages with durations (e.g. how long it took to transition to prepared and running states).
+  - `Async::Service::Configuration.build{|loader|...}` can now take an argument for more flexible configuration construction.
 
 ## v0.17.0